+++ components/Content/RoverData/RoverExplore.tsx
@@ -3,11 +3,11 @@ import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
 
 // UI elements
-import { Button } from "../../ui/Button";
-import { Checkbox } from "../../ui/checkbox";
+import { Button } from "../../Core/ui/Button";
+import { Checkbox } from "../../Core/ui/checkbox";
 // import { ResponsiveLine } from "@nivo/line";
-import { Input } from "../../ui/input";
-import { Textarea } from "../../ui/TextArea";
+import { Input } from "../../Core/ui/input";
+import { Textarea } from "../../Core/ui/TextArea";
 
 export default function RoverExploreDesktop({ id }: { id: string }) {
     const supabase = useSupabaseClient();
