+++ components/Core/ui/Avatar.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as AvatarPrimitive from '@radix-ui/react-avatar';
 
-import { cn } from "../../lib/uitls";
+import { cn } from "../../../lib/uitls";
 
 const Avatar = React.forwardRef<
 React.ElementRef<typeof AvatarPrimitive.Root>,
