+++ components/Core/ui/Card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../lib/uitls";
+import { cn } from "../../../lib/uitls";
 
 const Card = React.forwardRef<
 HTMLDivElement,
