+++ components/Core/ui/Form.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as LabelPrimitive from "@radix-ui/react-label";
 import { Slot } from "@radix-ui/react-slot";
 import { Controller, ControllerProps, FieldPath, FieldValues, FormProvider, useFormContext } from "react-hook-form";
-import { cn } from "../../lib/uitls";
+import { cn } from "../../../lib/uitls";
 import { Label } from "./Label";
 
 const Form = FormProvider;
